+++ docs/assets/modals.stories-CNXiDOvr.js
@@ -1,0 +1,10 @@
+import{_ as t}from"./UluButton-C2xVhAbG.js";import"./iframe-DEu6WBfK.js";import"./preload-helper-BJwshlQW.js";import"./UluAction-CAwkoBee.js";const p={title:"Plugins/modals"},e=()=>({components:{UluButton:t},template:`
+    <UluButton @click="$uluModals.open('demo')" text="Open App-wide Modal"/>
+  `}),o=e.bind({});o.args={};o.parameters={...o.parameters,docs:{...o.parameters?.docs,source:{originalSource:`() => ({
+  components: {
+    UluButton
+  },
+  "template": \`
+    <UluButton @click="$uluModals.open('demo')" text="Open App-wide Modal"/>
+  \`
+})`,...o.parameters?.docs?.source}}};const r=["GlobalApiTest"];export{o as GlobalApiTest,r as __namedExportsOrder,p as default};