+++ docs/assets/useScrollAnchorSections-DC9EkygC.js
@@ -1,0 +1,1 @@
+import{J as n}from"./iframe-DEu6WBfK.js";function r(){const o=n("uluScrollAnchorsSections");return o||console.warn("useScrollAnchorSections() must be used within an UluScrollAnchors component provider."),o}export{r as u};